requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pyyaml==6.0.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
chromadb==0.5.23
PyMuPDF==1.25.3
python-multipart==0.0.12
aiosqlite==0.20.0
